PiperOrigin-RevId: 464555461

+++ tensorflow/compiler/xla/mlir_hlo/include/mlir-hlo/Dialect/mhlo/transforms/map_mhlo_to_scalar_op.h
@@ -452,7 +452,7 @@ inline Value mapReducePrecisionOpToStdScalarOp(
   APInt expBitsMask(nbits, 1);
   expBitsMask = ((expBitsMask << srcExponentBits) - 1) << srcMantissaBits;
 
-  if (destMantissaBits < srcMantissaBits) {
+  if (destMantissaBits < static_cast<int>(srcMantissaBits)) {
     // Last remaining mantissa bit.
     APInt lastMantissaBitMask(nbits, 1);
     lastMantissaBitMask <<= srcMantissaBits - destMantissaBits;
